+++ taskcluster/rb_taskgraph/signing_apks.py
@@ -1,0 +1,30 @@
+# This Source Code Form is subject to the terms of the Mozilla Public
+# License, v. 2.0. If a copy of the MPL was not distributed with this
+# file, You can obtain one at http://mozilla.org/MPL/2.0/.
+"""
+Apply some defaults and minor modifications to the jobs defined in the build
+kind.
+"""
+
+from __future__ import absolute_import, print_function, unicode_literals
+
+from taskgraph.transforms.base import TransformSequence
+
+transforms = TransformSequence()
+
+
+@transforms.add
+def build_signing_task(config, tasks):
+    for task in tasks:
+        dep = task["primary-dependency"]
+        task["dependencies"] = {"build": dep.label}
+        task["worker"]["upstream-artifacts"] = [
+            {
+                "taskId": {"task-reference": "<build>"},
+                "taskType": "build",
+                "paths": dep.attributes["apks"].values(),
+                "formats": ["autograph_apk_reference_browser"],
+            }
+        ]
+        del task["primary-dependency"]
+        yield task
